refactor: update data types from int to uint16_t

+++ Pacman/include/entities/entity.h
@@ -3,6 +3,7 @@
 
 #include <stdbool.h>
 #include "../utils/vector2d.h"
+#include <stdint.h>
 
 // Forward declarations for polymorphism
 struct Entity;
@@ -23,14 +24,14 @@ typedef struct Entity
 {
     Vector2D position;       // Position of the entity
     Vector2D direction;      // Direction of movement
-    int speed;               // Movement speed
+    uint16_t speed;          // Movement speed
     EntityType type;         // Type of entity
     EntityUpdateFunc update; // Function pointer for updating the entity
     EntityRenderFunc render; // Function pointer for rendering the entity
 } Entity;
 
 // Function prototypes
-void entity_init(Entity *entity, EntityType type, Vector2D position, Vector2D direction, int speed);
+void entity_init(Entity *entity, EntityType type, Vector2D position, Vector2D direction, uint16_t speed);
 void entity_update(Entity *entity);
 void entity_render(const Entity *entity);
